Boss

+++ Assets/Entities/Scripts/HealthBar.cs
@@ -5,18 +5,18 @@ public class HealthBar : MonoBehaviour
 {
     public Image Bar;
     public PlayerManager player;
-    private int vidaBarra;
+    private float vidaBarra;
     
     // Start is called before the first frame update
     void Start()
     {
-        vidaBarra=player.vida/100;
+        vidaBarra=(float)player.vida/100;
     }
 
     // Update is called once per frame
     void Update()
     {
-        vidaBarra=player.vida/100;
+        vidaBarra=(float)player.vida/100;
         Bar.fillAmount=vidaBarra;
     }
 }
